Refactor imports from backtest
