+++ src/coin_test/backtest/__init__.py
@@ -1,0 +1,5 @@
+"""backtesting module of the coin-test library."""
+from .trade import Trade
+from .trade_request import TradeRequest
+
+__all__ = ["Trade", "TradeRequest"]
